# web/openapi.yaml
basePath: /api/v1
definitions:
...
    response.ErrorResponse:
        properties:
            error:
                type: string
            success:
                type: boolean
        type: object
...
    response.MediaInfo:
        properties:
            media_id:
                type: string
            thumbnail_url:
                type: string
            url:
                type: string
        type: object
...
    struktur.StrukturPaginatedResponse:
        properties:
            pagination:
                additionalProperties: true
                type: object
            struktur:
                items:
                    $ref: '#/definitions/struktur.StrukturResponse'
                type: array
        type: object
    struktur.StrukturResponse:
        properties:
            created_at:
                description: ISO 8601 format
                type: string
            description:
                type: string
            email:
                type: string
            id:
                type: string
            name:
                type: string
            phone:
                type: string
            position:
                type: string
            profile_image_url:
                description: 'Deprecated: use profile_media'
                type: string
            profile_media:
                $ref: '#/definitions/response.MediaInfo'
            updated_at:
                description: ISO 8601 format
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: REST API for the Desa (village) information system. Implements Clean Architecture with strict dependency rules, JWT (HS256) authentication, Casbin RBAC, PostgreSQL persistence, and rate limiting.
    title: Desa API
    version: "1.0"
paths:
...
    /public/struktur/{id}:
        get:
            description: Public full detail.
            parameters:
                - description: Struktur UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/struktur.StrukturResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: Get struktur (public)
            tags:
                - struktur
    /public/struktur/list:
        get:
            description: Public paginated list.
            parameters:
                - description: Page
                  in: query
                  name: page
                  type: integer
                - description: Page size
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/struktur.StrukturPaginatedResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: List struktur (public)
            tags:
                - struktur
...
    /struktur:
        get:
            description: 'RBAC: struktur:read.'
            parameters:
                - description: Search
                  in: query
                  name: q
                  type: string
                - description: Page
                  in: query
                  name: page
                  type: integer
                - description: Page size
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/struktur.StrukturPaginatedResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: List struktur (admin)
            tags:
                - struktur
        post:
            consumes:
                - multipart/form-data
            description: 'Multipart. Required: name. Optional: position, email, phone, description, profile_image. RBAC: struktur:write.'
            parameters:
                - description: Member name
                  in: formData
                  name: name
                  required: true
                  type: string
                - description: Position
                  in: formData
                  name: position
                  type: string
                - description: Email
                  in: formData
                  name: email
                  type: string
                - description: Phone
                  in: formData
                  name: phone
                  type: string
                - description: Description
                  in: formData
                  name: description
                  type: string
                - description: Profile image
                  in: formData
                  name: profile_image
                  type: file
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/struktur.StrukturResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "401":
                    description: Unauthorized
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "403":
                    description: Forbidden
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Create struktur member (admin)
            tags:
                - struktur
...
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'JWT bearer token. Format: `Bearer {token}`'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...